requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
python-jose
passlib[bcrypt]
sqlalchemy